requirements.txt
python-dotenv
eventlet~=0.24.1
opencv-python~=4.9.0.40
#flask==1.0.2
#flask-cors==3.0.7
#flask-login==0.4.1
#flask-restplus==0.12.1
#flask-mongoengine==0.9.5
#flask-socketio==3.3.2
fastapi~=0.104
fastapi-socketio
websockets
mongoengine~=0.28
numpy
cython
scikit-image
requests
google_images_download~=2.5.0
watchdog
pytest~=8.2.0
#pytest-ordering~=0.6.1
imantics~=0.1.9
celery~=5.4.0
Shapely~=2.0.4
scipy
Pillow
matplotlib
keras~=2.1.1
h5py
imgaug
python-jose
python-multipart
gunicorn~=22.0.0
pycocotools
werkzeug